+++ wdom/tests/remote_browser/test_tag.py
@@ -135,8 +135,8 @@ class TestInput(RemoteBrowserTestCase, TestCase):
         self.document.body.prepend(self.root)
         self.start()
 
-    @unittest.skipIf(os.environ.get('TRAVIS', False),
-                     reason='This test not pass only on travis')
+    # @unittest.skipIf(os.environ.get('TRAVIS', False),
+    #                  reason='This test not pass only on travis')
     def test_textinput(self):
         self.set_element(self.input)
         self.element.send_keys('a')
@@ -153,8 +153,8 @@ class TestInput(RemoteBrowserTestCase, TestCase):
             lambda: self.element.get_attribute('value') == 'ad')
         self.assertEqual(self.input.value, 'ad')
 
-    @unittest.skipIf(os.environ.get('TRAVIS', False),
-                     reason='This test not pass only on travis')
+    # @unittest.skipIf(os.environ.get('TRAVIS', False),
+    #                  reason='This test not pass only on travis')
     def test_textarea(self):
         self.set_element(self.textarea)
         self.element.send_keys('abc')
@@ -171,8 +171,8 @@ class TestInput(RemoteBrowserTestCase, TestCase):
             lambda: self.element.get_attribute('value') == 'abcdef')
         self.assertEqual(self.textarea.value, 'abcdef')
 
-    @unittest.skipIf(os.environ.get('TRAVIS', False),
-                     reason='This test not pass only on travis')
+    # @unittest.skipIf(os.environ.get('TRAVIS', False),
+    #                  reason='This test not pass only on travis')
     def test_checkbox(self):
         self.set_element(self.checkbox)
         self.element.click()
@@ -191,8 +191,8 @@ class TestInput(RemoteBrowserTestCase, TestCase):
         self.assertIsNone(self.element.get_attribute('checked'))
         self.assertIsFalse(self.checkbox.checked)
 
-    @unittest.skipIf(os.environ.get('TRAVIS', False),
-                     reason='This test not pass only on travis')
+    # @unittest.skipIf(os.environ.get('TRAVIS', False),
+    #                  reason='This test not pass only on travis')
     def test_checkbox_label(self):
         self.set_element(self.check_l)
         self.element.click()
@@ -203,8 +203,8 @@ class TestInput(RemoteBrowserTestCase, TestCase):
         self.wait_until(lambda: not self.checkbox.checked)
         self.assertFalse(self.checkbox.checked)
 
-    @unittest.skipIf(os.environ.get('TRAVIS', False),
-                     reason='This test not pass only on travis')
+    # @unittest.skipIf(os.environ.get('TRAVIS', False),
+    #                  reason='This test not pass only on travis')
     def test_radios(self):
         self.assertFalse(self.radio1.checked)
         self.assertFalse(self.radio2.checked)
@@ -231,8 +231,8 @@ class TestInput(RemoteBrowserTestCase, TestCase):
         self.assertTrue(self.radio2.checked)
         self.assertTrue(self.radio3.checked)
 
-    @unittest.skipIf(os.environ.get('TRAVIS', False),
-                     reason='This test not pass only on travis')
+    # @unittest.skipIf(os.environ.get('TRAVIS', False),
+    #                  reason='This test not pass only on travis')
     def test_radios_label(self):
         self.set_element(self.radio1_l)
         self.element.click()
@@ -266,8 +266,8 @@ class TestSelect(RemoteBrowserTestCase, TestCase):
         self.document.body.prepend(self.root)
         self.start()
 
-    @unittest.skipIf(os.environ.get('TRAVIS', False),
-                     reason='This test not pass only on travis')
+    # @unittest.skipIf(os.environ.get('TRAVIS', False),
+    #                  reason='This test not pass only on travis')
     def test_select(self):
         self.set_element(self.select)
         self.element.select_by_index(1)